requirements.txt
airtable-python-wrapper==0.15.3
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
base58==2.1.1
black==21.10b0
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
docutils==0.18
entrypoints==0.3
filelock==3.3.2
flake8==4.0.1
flit==3.4.0
flit-core==3.4.0
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
importlib-resources==5.4.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.18.0
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.21.4
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
py==1.11.0
pyarrow==6.0.0
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.7.1
pyflakes==2.4.0
Pygments==2.10.0
Pympler==0.9
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.2
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
streamlit==1.2.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tomli-w==0.4.0
toolz==0.11.2
tornado==6.1
tox==3.24.4
traitlets==5.1.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
virtualenv==20.10.0
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.6.0
